tweet/requirements.txt
absl-py==0.10.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
curl==0.0.1
cycler==0.10.0
Cython==0.29.22
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
dm-tree==0.1.5
entrypoints==0.3
fastapi==0.63.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-cloud-bigquery==2.11.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
httplib2==0.19.0
idna==2.10
importlib-resources==5.1.2
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.10
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
proto-plus==1.15.0
protobuf==3.15.5
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.8.1
pydot==1.4.2
Pygments==2.8.1
pypandoc==1.5
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.95
seqeval==1.2.2
six==1.15.0
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
tensorflow-metadata==0.28.0
tensorflow-model-optimization==0.5.0
tensorflow-text==2.4.3
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
tf-models-official==2.4.0
tf-slim==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typeguard==2.11.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
uvicorn==0.13.4
watermark==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
